+++ scripts/test_ingestion.py
@@ -7,7 +7,7 @@ from langchain.embeddings import OpenAIEmbeddings
 
 dotenv.load_dotenv()
 embeddings_key = os.getenv("API_KEY")
-docsearch = FAISS.load_local('outputs/inputs', OpenAIEmbeddings(openai_api_key=embeddings_key))
+docsearch = FAISS.load_local('outputs', OpenAIEmbeddings(openai_api_key=embeddings_key))
 
 d1 = docsearch.similarity_search("Whats new in 1.5.3?")
 print(d1)
